# [WIP] Improve loc performance by using join