--- a/python/cudf/cudf/dataframe/categorical.py
+++ b/python/cudf/cudf/dataframe/categorical.py
@@ -5,7 +5,6 @@ import pandas as pd
 import pyarrow as pa
 from pandas.core.dtypes.dtypes import CategoricalDtype
 
-import cudf
 import cudf.bindings.copying as cpp_copying
 import cudf.bindings.replace as cpp_replace
 from cudf.bindings.cudf_cpp import get_ctype_ptr