--- a/python/cudf/cudf/dataframe/numerical.py
+++ b/python/cudf/cudf/dataframe/numerical.py
@@ -117,14 +117,12 @@ class NumericalColumn(columnops.TypedColumnBase):
     def astype(self, dtype):
         from cudf.dataframe import datetime, string
 
-        if self.dtype == dtype:
+        if utils.dtype_equals(dtype, self.dtype):
             return self
 
-        elif dtype == np.dtype("object") or np.issubdtype(
-            dtype, np.dtype("U").type
-        ):
+        elif pd.api.types.is_string_dtype(dtype):
             if len(self) > 0:
-                if self.dtype in (np.dtype("int8"), np.dtype("int16")):
+                if utils.dtype_equals(dtype, ("int8", "int16")):
                     dev_array = self.astype("int32").data.mem
                 else:
                     dev_array = self.data.mem