Uses join (credit to @kkraus14) to improve `.loc` performance with non-contiguous row selection.

Before:

```
In [1]: import cudf                                 
In [2]: import numpy as np                          
In [3]: df = cudf.datasets.randomdata(100_000, {'a': float, 'b': int, 'c': str, 'd': 'category'})       
In [4]: i = np.random.choice(range(100_000), 200)   
In [5]: %timeit df.loc[i]                           
3.67 s ± 11.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

After:
```
In [1]: import cudf                                 
In [2]: import numpy as np                          
In [3]: df = cudf.datasets.randomdata(100_000, {'a': float, 'b': int, 'c': str, 'd': 'category'})       
In [4]: i = np.random.choice(range(100_000), 200)   
In [5]: %timeit df.loc[i]                           
40.6 ms ± 2.72 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```


This is WIP until #2221 is merged and I can figure out how to incorporate it,