@shwina 2019-07-31
Closing this PR in favour of #2387 - where I will include these changes